pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "google-genai==1.12.1",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "python-dotenv==1.1.0",
    "ruff>=0.15.1",
]
